After step 8, part 1 - Added PHPUnit test configuration and a test.
Added a test. You will need to install phpunit to run the test. I had
to adjust the configuration for the phpunit.xml.dist file to look for
tests in all directories in our application from what was in
http://fabien.potencier.org/article/57/create-your-own-framework-on-top-
of-the-symfony2-components-part-8. I’m including PHPUnit, but you can
also follow the instructions here to get installed properly in your
environment: http://phpunit.de/manual/current/en/installation.html. To
run the tests, go to this directory on the command line and type “php
phpunit.phar”. You will need to have a php installation available on
the command line for this to work.